Added IAsyncEnumerable wherever Task<IEnumerable>s were sold.
Removed IComparer for all repositories.

+++ Startitecture.Orm.SqlClient.Tests/JsonInsertTests.cs
@@ -10,6 +10,7 @@ namespace Startitecture.Orm.SqlClient.Tests
     using System.Collections.Generic;
     using System.Data;
     using System.Linq;
+    using System.Threading;
     using System.Threading.Tasks;
 
     using global::AutoMapper;
@@ -118,15 +119,18 @@ namespace Startitecture.Orm.SqlClient.Tests
 
                 // Delete the existing rows.
                 var fieldSelection = Query.Select<FieldRow>().Where(set => set.AreEqual(row => row.Name, "INS_%"));
-                var existingFields = fieldRepository.SelectEntities(fieldSelection);
+                var existingFields = fieldRepository.SelectEntities(fieldSelection).ToList();
 
-                fieldValueRepository.DeleteSelection(
-                    Query.Select<FieldValueRow>()
-                        .Where(set => set.Include(row => row.FieldId, existingFields.Select(field => field.FieldId).ToArray())));
+                if (existingFields.Any())
+                {
+                    fieldValueRepository.DeleteSelection(
+                        Query.Select<FieldValueRow>()
+                            .Where(set => set.Include(row => row.FieldId, existingFields.Select(field => field.FieldId).ToArray())));
 
-                fieldRepository.DeleteSelection(fieldSelection);
+                    fieldRepository.DeleteSelection(fieldSelection);
+                }
+
                 var fieldInsertCommand = new JsonInsert<FieldRow>(structuredCommandProvider, provider.DatabaseContext);
-
                 fieldInsertCommand.Execute(
                     fields.Select(
                         field => new FieldTableTypeRow
@@ -406,10 +410,11 @@ namespace Startitecture.Orm.SqlClient.Tests
                          };
 
             var providerFactory = new SqlClientProviderFactory(new DataAnnotationsDefinitionProvider());
+            var cancellationToken = CancellationToken.None;
 
             await using (var provider = providerFactory.Create(ConfigurationRoot.GetConnectionString("OrmTestDb")))
             {
-                var transaction = await provider.BeginTransactionAsync().ConfigureAwait(false);
+                var transaction = await provider.BeginTransactionAsync(cancellationToken).ConfigureAwait(false);
 
                 // Set up the structured command provider.
                 var structuredCommandProvider = new JsonParameterCommandFactory(provider.DatabaseContext);
@@ -418,14 +423,24 @@ namespace Startitecture.Orm.SqlClient.Tests
 
                 // Delete the existing rows.
                 var fieldSelection = Query.Select<FieldRow>().Where(set => set.AreEqual(row => row.Name, "INS_%"));
-                var existingFields = await fieldRepository.SelectEntitiesAsync(fieldSelection).ConfigureAwait(false);
+                var existingFields = new List<Field>();
 
-                await fieldValueRepository.DeleteSelectionAsync(
-                        Query.Select<FieldValueRow>()
-                            .Where(set => set.Include(row => row.FieldId, existingFields.Select(field => field.FieldId).ToArray())))
-                    .ConfigureAwait(false);
+                await foreach (var item in fieldRepository.SelectEntitiesAsync(fieldSelection, cancellationToken).ConfigureAwait(false))
+                {
+                    existingFields.Add(item);
+                }
 
-                await fieldRepository.DeleteSelectionAsync(fieldSelection).ConfigureAwait(false);
+                if (existingFields.Any())
+                {
+                    await fieldValueRepository.DeleteSelectionAsync(
+                            Query.Select<FieldValueRow>()
+                                .Where(set => set.Include(row => row.FieldId, existingFields.Select(field => field.FieldId).ToArray())),
+                            cancellationToken)
+                        .ConfigureAwait(false);
+
+                    await fieldRepository.DeleteSelectionAsync(fieldSelection, cancellationToken).ConfigureAwait(false);
+                }
+
                 var fieldInsertCommand = new JsonInsert<FieldRow>(structuredCommandProvider, provider.DatabaseContext);
 
                 await fieldInsertCommand.ExecuteAsync(
@@ -434,10 +449,11 @@ namespace Startitecture.Orm.SqlClient.Tests
                                      {
                                          Name = field.Name,
                                          Description = field.Description
-                                     }))
+                                     }),
+                        cancellationToken)
                     .ConfigureAwait(false);
 
-                await transaction.CommitAsync().ConfigureAwait(false);
+                await transaction.CommitAsync(cancellationToken).ConfigureAwait(false);
             }
         }
 
@@ -452,6 +468,7 @@ namespace Startitecture.Orm.SqlClient.Tests
         public async Task ExecuteWithIdentityUpdateAsync_JsonInsertForGenericSubmission_DoesNotThrowException()
         {
             var providerFactory = new SqlClientProviderFactory(new DataAnnotationsDefinitionProvider());
+            var cancellationToken = CancellationToken.None;
 
             await using (var provider = providerFactory.Create(ConfigurationRoot.GetConnectionString("OrmTestDb")))
             {
@@ -459,14 +476,17 @@ namespace Startitecture.Orm.SqlClient.Tests
 
                 var domainIdentity = await identityRepository.FirstOrDefaultAsync(
                                              Query.Select<DomainIdentity>()
-                                                 .Where(set => set.AreEqual(identity => identity.UniqueIdentifier, Environment.UserName)))
-                                         .ConfigureAwait(false) ?? await identityRepository.SaveAsync(
+                                                 .Where(set => set.AreEqual(identity => identity.UniqueIdentifier, Environment.UserName)),
+                                             cancellationToken)
+                                         .ConfigureAwait(false)
+                                     ?? await identityRepository.SaveAsync(
                                              new DomainIdentity(Environment.UserName)
                                              {
                                                  FirstName = "King",
                                                  MiddleName = "T.",
                                                  LastName = "Animal"
-                                             })
+                                             },
+                                             cancellationToken)
                                          .ConfigureAwait(false);
 
                 var expected = new GenericSubmission("My Submission", domainIdentity);
@@ -534,9 +554,15 @@ namespace Startitecture.Orm.SqlClient.Tests
                 // TODO: Return names only from the repo as a dynamic
                 var fields = expected.SubmissionValues.Select(value => value.Field).Distinct().ToDictionary(field => field.Name, field => field);
                 var inclusionValues = fields.Keys.ToArray();
-                var existingFields = await fieldRepository.SelectEntitiesAsync(
-                                             new EntitySelection<Field>().Where(set => set.Include(field => field.Name, inclusionValues)))
-                                         .ConfigureAwait(false);
+                var existingFields = new List<Field>();
+
+                await foreach (var item in fieldRepository.SelectEntitiesAsync(
+                                       new EntitySelection<Field>().Where(set => set.Include(field => field.Name, inclusionValues)),
+                                       cancellationToken)
+                                   .ConfigureAwait(false))
+                {
+                    existingFields.Add(item);
+                }
 
                 foreach (var field in existingFields)
                 {
@@ -546,13 +572,13 @@ namespace Startitecture.Orm.SqlClient.Tests
 
                 foreach (var field in fields.Values.Where(field => field.FieldId.HasValue == false))
                 {
-                    await fieldRepository.SaveAsync(field).ConfigureAwait(false);
+                    await fieldRepository.SaveAsync(field, cancellationToken).ConfigureAwait(false);
                 }
 
                 var submissionRepository = new EntityRepository<GenericSubmission, GenericSubmissionRow>(provider, this.mapper);
 
-                var transaction = await provider.BeginTransactionAsync().ConfigureAwait(false);
-                await submissionRepository.SaveAsync(expected).ConfigureAwait(false);
+                var transaction = await provider.BeginTransactionAsync(cancellationToken).ConfigureAwait(false);
+                await submissionRepository.SaveAsync(expected, cancellationToken).ConfigureAwait(false);
 
                 var submissionId = expected.GenericSubmissionId.GetValueOrDefault();
                 Assert.AreNotEqual(0, submissionId);
@@ -570,7 +596,12 @@ namespace Startitecture.Orm.SqlClient.Tests
                                         };
 
                 var valuesCommand = new JsonInsert<FieldValueRow>(commandProvider, provider.DatabaseContext).SelectFromInserted();
-                var insertedValues = (await valuesCommand.ExecuteForResultsAsync(valuesList).ConfigureAwait(false)).ToList();
+                var insertedValues = new List<FieldValueRow>();
+
+                await foreach (var item in valuesCommand.ExecuteForResultsAsync(valuesList, cancellationToken).ConfigureAwait(false))
+                {
+                    insertedValues.Add(item);
+                }
 
                 // Map back to the domain object.
                 foreach (var value in expected.SubmissionValues)
@@ -597,7 +628,12 @@ namespace Startitecture.Orm.SqlClient.Tests
 
                 // Reassign with our added identities
                 // TODO: create dictionary for seeks
-                var insertedElementRows = (await elementsCommand.ExecuteForResultsAsync(elementsList).ConfigureAwait(false)).ToList();
+                var insertedElementRows = new List<FieldValueElementRow>();
+
+                await foreach (var item in elementsCommand.ExecuteForResultsAsync(elementsList, cancellationToken).ConfigureAwait(false))
+                {
+                    insertedElementRows.Add(item);
+                }
 
                 foreach (var element in elementsList)
                 {
@@ -609,31 +645,34 @@ namespace Startitecture.Orm.SqlClient.Tests
                     .InsertInto(row => row.DateElementId, row => row.Value)
                     .From<FieldValueElementTableTypeRow>(row => row.FieldValueElementId, row => row.DateElement);
 
-                await dateElementsCommand.ExecuteAsync(elementsList.Where(row => row.DateElement.HasValue)).ConfigureAwait(false);
+                await dateElementsCommand.ExecuteAsync(elementsList.Where(row => row.DateElement.HasValue), cancellationToken).ConfigureAwait(false);
 
                 var floatElementsCommand = new JsonInsert<FloatElementRow>(commandProvider, provider.DatabaseContext)
                     .InsertInto(row => row.FloatElementId, row => row.Value)
                     .From<FieldValueElementTableTypeRow>(row => row.FieldValueElementId, row => row.FloatElement);
 
-                await floatElementsCommand.ExecuteAsync(elementsList.Where(row => row.FloatElement.HasValue)).ConfigureAwait(false);
+                await floatElementsCommand.ExecuteAsync(elementsList.Where(row => row.FloatElement.HasValue), cancellationToken)
+                    .ConfigureAwait(false);
 
                 var integerElementsCommand = new JsonInsert<IntegerElementRow>(commandProvider, provider.DatabaseContext)
                     .InsertInto(row => row.IntegerElementId, row => row.Value)
                     .From<FieldValueElementTableTypeRow>(row => row.FieldValueElementId, row => row.IntegerElement);
 
-                await integerElementsCommand.ExecuteAsync(elementsList.Where(row => row.IntegerElement.HasValue)).ConfigureAwait(false);
+                await integerElementsCommand.ExecuteAsync(elementsList.Where(row => row.IntegerElement.HasValue), cancellationToken)
+                    .ConfigureAwait(false);
 
                 var moneyElementsCommand = new JsonInsert<MoneyElementRow>(commandProvider, provider.DatabaseContext)
                     .InsertInto(row => row.MoneyElementId, row => row.Value)
                     .From<FieldValueElementTableTypeRow>(row => row.FieldValueElementId, row => row.MoneyElement);
 
-                await moneyElementsCommand.ExecuteAsync(elementsList.Where(row => row.MoneyElement.HasValue)).ConfigureAwait(false);
+                await moneyElementsCommand.ExecuteAsync(elementsList.Where(row => row.MoneyElement.HasValue), cancellationToken)
+                    .ConfigureAwait(false);
 
                 var textElementsCommand = new JsonInsert<TextElementRow>(commandProvider, provider.DatabaseContext)
                     .InsertInto(row => row.TextElementId, row => row.Value)
                     .From<FieldValueElementTableTypeRow>(row => row.FieldValueElementId, row => row.TextElement);
 
-                await textElementsCommand.ExecuteAsync(elementsList.Where(row => row.TextElement != null)).ConfigureAwait(false);
+                await textElementsCommand.ExecuteAsync(elementsList.Where(row => row.TextElement != null), cancellationToken).ConfigureAwait(false);
 
                 // Attach the values to the submission
                 var genericValueSubmissions = from v in insertedValues
@@ -644,8 +683,8 @@ namespace Startitecture.Orm.SqlClient.Tests
                                                      };
 
                 var submissionCommand = new JsonInsert<GenericSubmissionValueRow>(commandProvider, provider.DatabaseContext);
-                await submissionCommand.ExecuteAsync(genericValueSubmissions).ConfigureAwait(false);
-                await transaction.CommitAsync().ConfigureAwait(false);
+                await submissionCommand.ExecuteAsync(genericValueSubmissions, cancellationToken).ConfigureAwait(false);
+                await transaction.CommitAsync(cancellationToken).ConfigureAwait(false);
             }
         }
